+++ src/Symfony/Component/Messenger/Tests/Middleware/AllowNoHandlerMiddlewareTest.php
@@ -12,21 +12,23 @@
 namespace Symfony\Component\Messenger\Tests\Middleware;
 
 use PHPUnit\Framework\TestCase;
+use Symfony\Component\Messenger\Envelope;
 use Symfony\Component\Messenger\Exception\NoHandlerForMessageException;
 use Symfony\Component\Messenger\Middleware\AllowNoHandlerMiddleware;
 use Symfony\Component\Messenger\Tests\Fixtures\DummyMessage;
 
 class AllowNoHandlerMiddlewareTest extends TestCase
 {
-    public function testItCallsNextMiddlewareAndReturnsItsResult()
+    public function testItCallsNextMiddleware()
     {
         $message = new DummyMessage('Hey');
+        $envelope = new Envelope($message);
 
         $next = $this->createPartialMock(\stdClass::class, array('__invoke'));
-        $next->expects($this->once())->method('__invoke')->with($message)->willReturn('Foo');
+        $next->expects($this->once())->method('__invoke')->with($envelope);
 
         $middleware = new AllowNoHandlerMiddleware();
-        $this->assertSame('Foo', $middleware->handle($message, $next));
+        $middleware->handle($envelope, $next);
     }
 
     public function testItCatchesTheNoHandlerException()
@@ -36,7 +38,7 @@ class AllowNoHandlerMiddlewareTest extends TestCase
 
         $middleware = new AllowNoHandlerMiddleware();
 
-        $this->assertNull($middleware->handle(new DummyMessage('Hey'), $next));
+        $this->assertNull($middleware->handle(new Envelope(new DummyMessage('Hey')), $next));
     }
 
     /**
@@ -49,6 +51,6 @@ class AllowNoHandlerMiddlewareTest extends TestCase
         $next->expects($this->once())->method('__invoke')->will($this->throwException(new \RuntimeException('Something went wrong.')));
 
         $middleware = new AllowNoHandlerMiddleware();
-        $middleware->handle(new DummyMessage('Hey'), $next);
+        $middleware->handle(new Envelope(new DummyMessage('Hey')), $next);
     }
 }