+++ src/Symfony/Component/Messenger/Tests/Middleware/Enhancers/TraceableMiddlewareTest.php
@@ -26,14 +26,14 @@ class TraceableMiddlewareTest extends TestCase
     public function testHandle()
     {
         $busId = 'command_bus';
-        $envelope = Envelope::wrap($message = new DummyMessage('Hello'));
+        $envelope = new Envelope($message = new DummyMessage('Hello'));
 
         $middleware = $this->getMockBuilder(MiddlewareInterface::class)->getMock();
         $middleware->expects($this->once())
             ->method('handle')
-            ->with($message, $this->anything())
-            ->will($this->returnCallback(function ($message, callable $next) {
-                return $next($message);
+            ->with($envelope, $this->anything())
+            ->will($this->returnCallback(function ($envelope, callable $next) {
+                $next($envelope);
             }))
         ;
 
@@ -41,8 +41,7 @@ class TraceableMiddlewareTest extends TestCase
         $next
             ->expects($this->once())
             ->method('__invoke')
-            ->with($message)
-            ->willReturn($expectedReturnedValue = 'Hello')
+            ->with($envelope)
         ;
 
         $stopwatch = $this->createMock(Stopwatch::class);
@@ -58,7 +57,7 @@ class TraceableMiddlewareTest extends TestCase
 
         $traced = new TraceableMiddleware($middleware, $stopwatch, $busId);
 
-        $this->assertSame($expectedReturnedValue, $traced->handle($envelope, $next));
+        $traced->handle($envelope, $next);
     }
 
     /**
@@ -68,14 +67,14 @@ class TraceableMiddlewareTest extends TestCase
     public function testHandleWithException()
     {
         $busId = 'command_bus';
-        $envelope = Envelope::wrap($message = new DummyMessage('Hello'));
+        $envelope = new Envelope($message = new DummyMessage('Hello'));
 
         $middleware = $this->getMockBuilder(MiddlewareInterface::class)->getMock();
         $middleware->expects($this->once())
             ->method('handle')
-            ->with($message, $this->anything())
-            ->will($this->returnCallback(function ($message, callable $next) {
-                return $next($message);
+            ->with($envelope, $this->anything())
+            ->will($this->returnCallback(function ($envelope, callable $next) {
+                $next($envelope);
             }))
         ;