+++ src/Symfony/Component/Messenger/Tests/Middleware/ValidationMiddlewareTest.php
@@ -24,7 +24,7 @@ class ValidationMiddlewareTest extends TestCase
     public function testValidateAndNextMiddleware()
     {
         $message = new DummyMessage('Hey');
-        $envelope = Envelope::wrap($message);
+        $envelope = new Envelope($message);
 
         $validator = $this->createMock(ValidatorInterface::class);
         $validator
@@ -38,17 +38,15 @@ class ValidationMiddlewareTest extends TestCase
             ->expects($this->once())
             ->method('__invoke')
             ->with($envelope)
-            ->willReturn('Hello')
         ;
 
-        $result = (new ValidationMiddleware($validator))->handle($envelope, $next);
-
-        $this->assertSame('Hello', $result);
+        (new ValidationMiddleware($validator))->handle($envelope, $next);
     }
 
     public function testValidateWithStampAndNextMiddleware()
     {
-        $envelope = Envelope::wrap($message = new DummyMessage('Hey'))->with(new ValidationStamp($groups = array('Default', 'Extra')));
+        $message = new DummyMessage('Hey');
+        $envelope = (new Envelope($message))->with(new ValidationStamp($groups = array('Default', 'Extra')));
         $validator = $this->createMock(ValidatorInterface::class);
         $validator
             ->expects($this->once())
@@ -61,12 +59,9 @@ class ValidationMiddlewareTest extends TestCase
             ->expects($this->once())
             ->method('__invoke')
             ->with($envelope)
-            ->willReturn('Hello')
         ;
 
-        $result = (new ValidationMiddleware($validator))->handle($envelope, $next);
-
-        $this->assertSame('Hello', $result);
+        (new ValidationMiddleware($validator))->handle($envelope, $next);
     }
 
     /**
@@ -76,7 +71,7 @@ class ValidationMiddlewareTest extends TestCase
     public function testValidationFailedException()
     {
         $message = new DummyMessage('Hey');
-        $envelope = Envelope::wrap($message);
+        $envelope = new Envelope($message);
 
         $violationList = $this->createMock(ConstraintViolationListInterface::class);
         $violationList